Fix inconsistencies in distance checking when opening chests with mouse click and with space
The way it used to be was that chests can be opened with a mouse from 512 map units away, but with space only from 256 units away.

+++ Engine/MM7.h
@@ -3,10 +3,10 @@
 
 typedef unsigned int uint;
 
-#define PID(type, id) (uint32_t)((((8 * (id))) | (type)) & 0xFFFF)  // packed id
+#define PID(type, id) (uint16_t)((((8 * (id))) | (type)) & 0xFFFF)  // packed id
 #define PID_TYPE(pid) (ObjectType)((pid)&7)          // extract type
 #define PID_ID(pid) (uint32_t)(((pid)&0xFFFF) >> 3)  // extract value
-#define PID_INVALID (-1)
+#define PID_INVALID (uint16_t)(-1)
 
 // typedef char _UNKNOWN;
 typedef unsigned int uint;
